Adds refactoring for all tests to be included in the custom DC shared tests.
