+++ server/webdriver/cdc_tests/autopush/vis_map_test.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Google LLC
+# Copyright 2024 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,10 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from server.webdriver.cdc_tests.autopush.cdc_base_webdriver import \
+    CdcAutopushTestBase
+from server.webdriver.shared_tests.vis_map_test import VisMapTestMixin
 
-from server.webdriver.base_dc_webdriver import BaseDcWebdriverTest
-from server.webdriver.shared_tests.vis_scatter_test import VisScatterTestMixin
 
-
-class TestVisMap(VisScatterTestMixin, BaseDcWebdriverTest):
-  """Class to test scatter visualization page. Tests come from VisScatterTestMixin."""
+class TestVisMap(VisMapTestMixin, CdcAutopushTestBase):
+  """Class to test the map visualization page for Custom DC. Tests come from VisMapTestMixin."""
