+++ server/webdriver/tests/event_page_test.py
@@ -28,8 +28,8 @@ FIRE_EVENT_DCID = 'fire/imsr0003Fire20152836427'
 DROUGHT_EVENT_DCID = 'stormEvent/nws5512667'
 
 
-# Class to test Event Pages.
 class TestEventPage(WebdriverBaseTest):
+  """Class to test Event Pages."""
 
   def test_page_cyclone(self):
     """Test a cyclone event page can be loaded successfully"""
@@ -44,7 +44,7 @@ class TestEventPage(WebdriverBaseTest):
 
     # Assert page title is correct
     WebDriverWait(self.driver, self.TIMEOUT_SEC).until(
-        EC.title_contains('Nicole - Event Page - Data Commons'))
+        EC.title_contains('Nicole - Event Page - ' + self.dc_title_string))
 
     # Check header section
     element_present = EC.presence_of_element_located((By.TAG_NAME, 'h1'))
@@ -177,7 +177,8 @@ class TestEventPage(WebdriverBaseTest):
 
     # Assert page title is correct
     WebDriverWait(self.driver, self.TIMEOUT_SEC).until(
-        EC.title_contains('stormEvent/nws5512667 - Event Page - Data Commons'))
+        EC.title_contains('stormEvent/nws5512667 - Event Page - ' +
+                          self.dc_title_string))
 
     # Check header section
     element_present = EC.presence_of_element_located((By.TAG_NAME, 'h1'))
