+++ server/webdriver/tests/homepage_test.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
@@ -19,9 +20,8 @@ from server.webdriver.base_dc_webdriver import BaseDcWebdriverTest
 from server.webdriver.shared_tests.homepage_test import HomepageTestMixin
 
 
-class TestPlaceLanding(HomepageTestMixin, BaseDcWebdriverTest):
+class TestHomepage(HomepageTestMixin, BaseDcWebdriverTest):
   """Tests for Homepage. Some tests come from HomepageTestMixin."""
-  dc_title_string = "Data Commons"
 
   def test_homepage_en_by_css(self):
     """Test homepage in EN."""
@@ -29,7 +29,7 @@ class TestPlaceLanding(HomepageTestMixin, BaseDcWebdriverTest):
     self.driver.get(self.url_ + '/')
 
     title_present = EC.text_to_be_present_in_element(
-        (By.CSS_SELECTOR, '#main-nav .navbar-brand'), 'Data Commons')
+        (By.CSS_SELECTOR, '#main-nav .navbar-brand'), self.dc_title_string)
     WebDriverWait(self.driver, self.TIMEOUT_SEC).until(title_present)
 
     hero_msg = self.driver.find_elements(By.ID, 'hero')[0]
@@ -43,7 +43,7 @@ class TestPlaceLanding(HomepageTestMixin, BaseDcWebdriverTest):
     self.driver.get(self.url_ + '/?hl=it')
 
     title_present = EC.text_to_be_present_in_element(
-        (By.CSS_SELECTOR, '#main-nav .navbar-brand'), 'Data Commons')
+        (By.CSS_SELECTOR, '#main-nav .navbar-brand'), self.dc_title_string)
     WebDriverWait(self.driver, self.TIMEOUT_SEC).until(title_present)
 
     hero_msg = self.driver.find_elements(By.ID, 'hero')[0]
